cityclean/requirements.txt
absl-py==2.0.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.7
click==8.0.3
comm==0.1.4
cryptography==41.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.59.0
django-bootstrap4==23.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
Flask==2.0.2
flatbuffers==23.5.26
fonttools==4.33.3
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.3
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.13.1
kiwisolver==1.4.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mssql-django==1.3
mysql-connector-python==8.2.0
mysqlclient==2.0.3
nest-asyncio==1.5.8
numpy==1.24.3
oauthlib==3.2.2
openai==0.19.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandas-stubs==1.2.0.35
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.0.1
pyparsing==3.0.9
pysmile==2.0.10
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
pyzmq==25.1.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
toml==0.10.2
tornado==6.3.3
tqdm==4.62.3
traitlets==5.11.1
typing_extensions==4.5.0
urllib3==1.26.7
wcwidth==0.2.8
Werkzeug==2.0.2
wrapt==1.15.0
